Block iPhone 5 and have to be in Database
When private_to_all_devices == "ON": iphone 5 / 4S /iPad 3 can be blocked when they are in database. Initially they are not blocked.
When clients_must_be_in_database == true: iPhone 5 / 4S / iPad 3 must be in database if private_to_all_devices is also "ON". That means that icloud contact sync must be enabled to establish a connection. iPhone 5 etc devices are automaticaly unlocked and have a new database-row in order to facilitate the connection but they can be blocked afterwards.